pyproject.toml
[project]
name = "sublang"
version = "0.1.0"
description = "Chatbot for software design in well-defined sublanguages"
readme = "README.md"
authors = [
    { name = "Jinglei Ren", email = "jinglei@merico.ai" }
]
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "litellm>=1.40.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
sublang = "sublang.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "twine>=6.1.0",
]
